move the config
