+++ geocode/geocode.go
@@ -5,7 +5,7 @@ import (
 	"log"
 )
 
-func GetCordinatesForAddress(address string, geo *geo.GoogleGeo) (lat float64, lon float64) {
+func GetCoordinates(address string, geo *geo.GoogleGeo) (lat float64, lon float64) {
 	res, err := geo.Geocode(address)
 	if err != nil {
 		log.Println(err)
